[Add],[Del] 코드 수정 및 삭제
view, viewcontroller 분리 및 코드 간편화
