+++ StudyGaemi/Views/BaseViewController.swift
@@ -74,4 +74,5 @@ extension UITextField {
         animation.repeatCount = 8
         self.layer.add(animation, forKey: "shake")
     }
+    
 }
